Contains schemamigration!
This commit allows assistants to give feedback for submissions in a way
that does not override the feedback received from the grader (though that
is still possible to do, if desired).

Viewing submissions (whether in inspect mode or from students view) will
show two feedbacks (assistant and grader) if available, or notice that
there is not feedback available.

Schemamigration adds assistant_feedback field to submissions