CONTAINS SCHEMAMIGRATION
Adds the possibility for assistants to give feedback about a
submission, without overriding grader generated feedback (though
this is still possible, if desired).

When viewing submission (either from student's view or from
assistants inspect view) feedbacks from assistant and grader are
shown separately. If feedback of that type is not available, info
is shown.