Change how permission denied message is handled
Before only the message given to error_msg was added to django contrib message
subsystem and thus visible on login and forbidden pages. Now, all permission
errors are added to that subsystem so they are visible and thus the behavior
is consistent between different permission classes.

In addition error_msg now supports set of extra features when working with
translated strings.

Useless code and code needed before django 1.10 is removed.