Fix changing value of unsubmitted ae-input
In the case of active element outputs dependent on multiple input elements,
it was possible to modify one input and submit another input so that
the evaluation of the output related to the submitted input would use
the other, unsubmitted, input value. This would create a conflict between
the content of the unsubmitted yet changed input value and other outputs
(not dependent on the input that actually as submitted) that would still
reflect the old input value.

To fix this the original input values from page load/input form submission
are saved, and these values are used in evaluation for inputs others than
the one that is being submitted by the user. Thus modifications to inputs
that are not submitted will be ignored and not saved.