Fix submitter form when old submitter was not in the list
Chrome and Firefox validate the input and if there is required fields it will complain that those are not filled.
We updated the selection after above check was done and thus browser didn't get there.

The selection system should be implemented in a bit different way (note the comment in the source).