Add LTI 1.1 support.
LTI Tool Providers may now post grades back to A+ using the Basic Outcomes Service.

Invalidate the points cache when the submitters of a submission are modified or added.
This is done by adding a signal handler for m2m_changed that behaves
similarly to the submission post save/delete signal.
(Many-to-many fields behave a little differently and do not trigger
the post save signal.)

Disable the regrade button for LTI exercises in the inspect submission page.

Update language locale strings. Add one new string.