Show exercises in correct order on all results page
I.e. in the order they appear in the forest.

Assume that anything interesting is a leaf. It should be relatively simple
to modify this to inspect the whole forest if needed in the future.

It should be possible to implement this also using search_exercises
which returns the non-root nodes of the forest as a list. That might lead
to a simpler solution code-wise but I feel tree traversal is the more intuitive
solution.