Fix issues in the course instance view content to setting
Particularly, the enrolled students value of 'view content to'
was buggy and allowed unenrolled users to see the course front
page and the structure of the course.

This commit adds a new enrollment page that is always accessible
to logged-in users and in public courses (view access to public)
also anonymous users may open the enrollment page.
The enrollment page is particularly needed when users may not
open the course front page before enrolling, in which case
logged-in users in the enrollment audience may enroll via
the enrollment page. Enrolling is still possible in the course
front page as well, which works well if the 'view access to'
value is anything but the strict "enrolled students".

In addition to course visibility, this commit fixes the
visibility of exercises. Enrollment exercises were also
especially checked since they are accessed while the user
has not enrolled.

The exercise overlay (which informs the user why they can not
submit to the exercise) now shows a link to the enrollment page
if the user is not enrolled and may not enroll. The enrollment
page shows the reason why they may not enroll. If they may
enroll, the enrollment button is directly in the overlay
as before.