Improve validation of course timings and prevent submissions after archiving
Students may not submit to exercises after the course archive time even if
the course module is still open. The archive time is a strict boundary.
Likewise, students may not ever open exercise model solutions after
the course lifesupport time.

If course lifesupport time and archive time are not set, they use a default
value that is one year later than the course ending time. The default value
is not stored in the database. Instead, the database value may be null and
the default is then dynamically computed from the course ending time.

Improved validation:
- Lifesupport time and archive time may be equal to each other and
  the course ending time. They still may not be earlier than the course ending
  time. Archive time may still not be earlier than lifesupport time.
- Course module opening time must be earlier than the closing time.
- Course module late submission deadline must be later than the closing time
  (if late submissions are enabled).