Remove unofficial submissions from the visualizations overview
Quick fix: remove unofficial submissions from the visualizations
page overview. The graphs and tables as well as CSV file downloads
do not include unofficial points now.

In the future, the API used by visualizations should maybe be
based on the points cache. Currently, it queries the database and
it does not check the confirm the level setting, which means
that visualizations may show unconfirmed points.
Not many courses use the confirm the level setting, though.