Fix a bug in the API: view.get_access_mode() is not defined
Commit 0447dc5a94c9ec71a469a30013df1f31294d3bec missed this part.

Some API view classes use the class CourseInstanceBaseMixin that
uses the permission class CourseVisiblePermission. That permission
calls the method get_access_mode() on the view object.
Access modes are not really supposed to be used in the API, but
the CourseVisiblePermission requires it. The AccessModePermission
that really checks the access mode is probably not used in the API
so the access mode value is not used outside the
CourseVisiblePermission.

CourseInstanceBaseMixin assumes that its super class defines
the method get_access_mode(). In the API, that was not the case
previously, which resulted in crashes. This commit fixes this.