Combine data from previous enrollment exercise submissions
When an enrollment exercise is opened, read all enrollment
exercise submissions of the user from the database.
Their submission data is combined.
It is then possible to prefill enrollment questionnaire
forms on other courses if they use a similar questionnaire as
a course that the student has already enrolled in.
The enrollment questionnaire form may be pre-filled by
frontend JavaScript. The backend embeds the necessary submission
data in the enrollment exercise page as a JSON dump.

The latest enrollment submission data available in the
HTML template is restricted so that it only contains form keys
that are also used by the exercise to be opened
(based on exercise_info.form_spec).