Fix frontend JavaScript bugs related to user tags.
Student cache was changed in
https://github.com/Aalto-LeTech/a-plus/commit/669ab061c14e97102ec6cca2faf38451f5a9f8a7#diff-1b1e1c4baf6ecacc04fb1210e6ab0f33L32
Some JS still used the old tag_ids variable. JS now uses tag_slugs
as intended (internal and external user tags have only slugs,
no ids). This fixes filtering of students by tag and the removal
of tags in the participants page.

In addition, if a tag is added to many students at once in the
participants page, the tag badge is added to each affected student
row in the table instead of adding it multiple times to a single
row.