Fix JS bug in the removal of user taggings
Previously, deleting user taggings in the course participants page
could fail because the JS code did not handle paginated responses
from the API. The code retrieved all user taggings for the tag
even though it was only supposed to remove one tagging.

The commit adds a new endpoint to the API so that it is easy to
delete taggings without retrieving all taggings (which was
previously done to find out the ID of the tagging to delete).
The new delete endpoint uses the same URL as the list view of
the user taggings viewset (since it may delete multiple taggings).

The new JS code for the deletion may also delete multiple taggings
in a single HTTP DELETE request.

The idea for the API router hack for adding DELETE to the list view
came from Jaakko.

Co-authored-by: Jaakko Kantojärvi <jaakko.kantojarvi@aalto.fi>