Fix inconsistencies in the ordinal_number parameter
The ordinal_number parameter of the grader protocol is a counter
for submission attempts. It is used in at least personalized and
randomized exercises in order to change the exercise after the
user has submitted.

The ordinal_number counter does not include rejected submissions
or submissions in the error state when a submission is uploaded
for grading. However, previously those submissions were included
in the counter when the exercise description was retrieved from
the exercise service. This commit fixes this inconsistency and
now rejected and errorneous submissions are excluded from the
counter when the exercise description is retrieved.
It is important that the counter behaves in the same way for
both cases (GET and POST to the exercise service) because
otherwise, a personalized exercise could show a different instance
to the user than what actually gets graded when the user submits.

The new behaviour also matches how A+ uses the submission limit
of the exercise (max submissions). That limit ignores rejected and
errorneous submissions.