Modify links so that they work
Links to nested chapters have modified keys and URLs of the links must have
similar modification. In multilingual courses, language postfixes are removed
from file names, when index files are combined.

This functionality was previously implemented in RST-tools, but for links
to work in the build folder of the course, some of the changes had to
be moved to A-plus code.