Enrollment exercises ignore the module close time
If a course has an enrollment exercise (questionnaire),
the exercise can be answered even if the module containing
the exercise has closed. The course enrollment ending time
still limits new enrollments.

Additionally, add a dot in the end of a sentence in
CourseModulePermission.

Add unit tests that submit to enrollment exercises after
the module closing time.