Confirm the level exercises need to be submitted.
Previously, all exercises that had 'points_to_pass' as 0 had the 'passed'
value 'True' as a default, even without any submissions. If the chapter
had an exercise with 'confirm_the_level' as 'True', it was often considered
passed and it didn't have to be submitted. By changing the default 'passed'
value for level confirming exercises, those exercises really have to be
submitted to get any points for the chapter.