Show initialized and waiting submissions in the submission count.
Even though the submission with statuses `initialized` or `waiting`
can be rejected, it is better to have them counted in the submission
count. It is more clear for the students so they can see how many
submissions they've made. If the submission gets rejected, it will be
removed from the total submission count.