Modules can be opened for reading before opening the exercises.
The reading opening time can be set before the actual opening time
to allow students to see the material, but not exercises. When not
set, the reading material is not opened before the exercises, as
previously.

This commit also modifies the HTML templates so that the links
to the chapters that are opened for reading are active, and links to
exercises in those chapters are disabled.

Dates shown in schedules use the opening time of the exercises.

Some unit tests have been included.