Do not collapse read-open modules in the exercise results page
If a module is open for reading, it is now opened by default in the
exercise results page (i.e., the HTML div element is not collapsed).
Add a badge to inform that this module is for reading only.
Add correct prepositions with dates (Opens on saturday).
Template tag `module_accessible` fixed to work properly, even when
opening time is not set.

Fixes #473