Separate chapters and exercises in the Edit course page
Texts on links for editing and viewing exercises now separate
content chapters and exercises.
Teachers should be able to more easily distinguish exercises
and chapters in the content tab of the edit course page.
Additionally, the view submissions link is now again visible for
exercises. It disappeared due to a bug and it has now been fixed.