Use AJAX in search fields
For example, in the edit course form, users are now
searched from the server API via AJAX in the
"add assistants" field. Previously, all users were
downloaded in the form HTML and users were searched
locally from the hidden option fields of a select element.
Downloading all users in the HTML significantly
increased the page size and slowed down the page
load time.

This commit adds the JavaScript code for such general
AJAX search form fields as well as some utilities.
In addition, the edit course form is modified so that
the add assistants field uses this AJAX search feature.

Grant teachers the permission to retrieve all users and
their basic information from the API. This is necessary
so that teachers may use AJAX search forms
(in order to add course assistants; user API is only
needed when the form handles users).

The new teacher permission class is implemented
inefficiently and should be improved in the future.