Do not require full_name on UserBriefSerializer
If it is required, it breaks the old JavaScript code for adding
new user taggings in the course participants page.
HTTP POST to /api/v2/courses/1/taggings/ received the response
"400 bad request":
{"user":{"full_name":["This field is required."]}}

The full_name field was added in December 2019:
https://github.com/apluslms/a-plus/commit/061af2a1cc5d9c8963d6e751720ad503621a639a#diff-a9270613245f22f0442abbb5cc326803

The full_name field should perhaps be read-only since it is based
on the user's first and last name fields.