Slightly improve the performance of IsTeacherOrAdminOrSelf
Check super users first so that they can pass without querying
all courses from the database.

Query Courses from the database instead of CourseInstances
because there are less Courses and at any rate, teachers are
defined in the Course.

Use a generator expression with the `any` function so that
the evaluation may stop at the first positive match.
It is enough to know that the user is a teacher on one course.
The previous list comprehension always evaluated `is_teacher()`
for all courses.