Remove user detail API queries from user search forms
In forms that use an AJAX-based user search field, the hidden
select element now contains data (names, student IDs and emails)
of the initially selected users. The data is used to render the
initial selection in ajax_search_select.js without querying
the API for the user details. In other words, the extra AJAX calls
to the API for each initially selected user have been removed.

The form classes must now set the initial values with
a UserProfile queryset instead of manually mapping the profiles to
user IDs. Such mapping was an ugly hack and now it is not needed.