Add initial_queryset to the SearchSelectField
Initial_queryset is used to render the choices of the HTML widget
(select multiple element). Previously, the queryset parameter was
used for that. Now, queryset is used for validating user input and
the separate initial_queryset defines the options rendered to the
HTML widget. We can now use the normal validation of the parent
class ModelMultipleChoiceField while still limiting the number of
choices rendered in the HTML widget. In summary, SearchSelectField
does not need to override the clean method for validation since
the parent class takes care of that with the queryset, and
SearchSelectField has a new initial_queryset parameter for defining
the rendered options of the HTML widget so that the widget may
render much fewer options than all valid choices of the queryset.

Update the form classes that use UsersSearchSelectField to use
the new API of the field class.