Add API endpoint for normal exercise submissions from students
The new API endpoint can be used by students. Previously, there
were endpoints related to submissions, but none of them allowed
students to submit new solutions for grading. The new endpoint
reads submission data from the HTTP POST request data and files
may be uploaded as well.

The URL is like this: `/api/v2/exercises/ID/submissions/submit/`.
ExerciseSubmissionsViewSet already had a create method that
only allows teachers to create new graded submissions; hence,
the new endpoint was added as an additional action in the viewset.
The new action method `submit` is very similar to the old post
method of the ExerciseView of the web app (normal non-API code).
However, the new endpoint is implemented with the Django REST
framework like everything else in the API and it supports the
token authentication that has always been enabled in the API.

The Django messages framework does not work well with the API
since it is tied to user sessions (in the browser). The old
BaseExercise.grade method is changed so that it no longer uses
the messages framework, but it returns the error message strings
in the errors attribute of the ExercisePage object. The errros
attribute existed before but it was never used.

API requests should include the `_aplus_group` parameter with the
value zero if the user submits alone. Otherwise, A+ uses the
group that the user has selected in the course level in the
A+ course page. For a group submission, the parameter defines
the ID of the student group. The user must be a member of the
group or else the group is ignored and the submission is made
alone.

The POST data must use the correct keys for the submitted data.
The keys are defined by the exercise and can be retrieved from
the API: `http://localhost:8000/api/v2/exercises/ID/`
(`form_spec` under `exercise_info`).

Finally, here is an example of a simple `curl` command for
submitting a new solution to the exercise with ID 21.
The submission is made alone and includes the file `solution.py`.

```
curl -v -X POST \
'http://localhost:8000/api/v2/exercises/21/submissions/submit/' \
-H 'Authorization: Token 94ad2abfc128a3da3b0541fbfea981bc62becd56' \
-F "_aplus_group=0" -F 'file1=@solution.py'
```