Create "skip links" for keyboard users
Adds two skip links - one at the top of the page which jumps
to the page contents, and another at the top of the "course menu"
which jumps to the course content.

It also adds matching `aria-label` attributes, to give the user
a picture of where they are in a consistent manner.

Fixes https://github.com/apluslms/a-plus/issues/559