Frontpage-sections for my and all courses
Commit divides the front page into two sections based on the
course instances of the user. Upper-section shows the courses a user
teaches or assists or is enrolled in, if any, and bottom-section
shows all the remaining ongoing courses.

A separate html-template was made for the common part of these
two sections.

Partly fixes: https://github.com/apluslms/a-plus/issues/425