Implement a consistent focus state with a polyfill
Previously we haven't had a single style for whatever has focus. This
adds a style to the most common components using a yellow colour that
isn't widely used at the moment.

We're using `!important` here as we want to ensure that whenever there
is a conflict, the focus state always wins. Since this is currently
implemented using a JS polyfill, there isn't a better, reliable way to
acheive this without needing CSS changes for every interface change.

In the future, I'm hopeful that `:focus-visible` becoming standard, and
a possible move to SASS, or another preprocessor, will enable a less
hacky-feeling solution here.

Fixes https://github.com/apluslms/a-plus/issues/546