Add default-messages for frontpage
In case there are no courses where the user is a
teacher, student or an assistant, the My courses -panel
is shown but with a default text of not having any active
courses.

Similarly, if there are no ongoing courses at all,
a default text indicating that there are no active courses
is shown.

Partly fixes: https://github.com/apluslms/a-plus/issues/425