Remove folding of the course news -content
Previously all the content-text of course news
was folded and only visible with a click of a
button. Now the folding has been removed and
the appearance of the news items modified so that
news would be more easily noticeable.

Also course news folding -button is made more
noticeable and accessible.