Save language to userprofile
When logged-in user changes language in their profile page, save it to
the language field in their userprofile and use it as default language
for displaying content instead of saving the language to a cookie. For
users who are not logged in, language is still saved to a cookie when
changed on course page.

1/3 of issue https://github.com/apluslms/a-plus/issues/518