Use card layout on course archive page
This switches the course archive to use a layout similar to the front
page, and adapts the frontpage `card` component to include a "condensed"
version.

The courses are arranged by starting calendar year.

This leads to more consistency and resolves the small accessibility
issues found in the page.

The link to the archive also moves from hidden inside a sentence to
alongside the section header on the front page, and renames the
card-related classes to be a little more generic.

Fixes https://github.com/apluslms/a-plus/issues/626