Change language based on url query parameter
Change language based on hl query parameter, for example ?hl=fi at the
end of the url shows the content in Finnish if the page is available in
that language. Otherwise, an error is raised.
* If the error is raised on other than course page, a 404 html page is
shown with a link to the material in default language.
* If the error is raised on course page, a 404 html page is shown with
a list of available languages and links for each.
* Language parameter stays in the url when navigating through page,
except when user changes language either on course page or on profile
page.