Implement new `.ap-button` component
This creates a new, custom button component.

The size matches the Bootstrap sizing, but the styles are a little
different, focusing on consistency and readability, in accordance with
the Button Guidelines, which are in "beta" in the design system:
https://apluslms.github.io/a-plus-design-system/components/button.html

Note that this commit doesn't take them into use, but simply makes them
available. It starts work on https://github.com/apluslms/a-plus/issues/601