Add points indicator component
Creates a new component with the associated view. The component isn't
used anywhere yet.

The component displays an SVG icon alongside text to show progress on a
question, chapter or course. The icon is coloured using CSS, and the
data provided to the partial view are used to decide which icon/words to
show.