Hotfix: Add accessibility statement
This adds an accessibility statement, according to the EU Accessibility
directive, and a template from Aalto University.

Since the statement should be specific to the institution, this is split
into the "local part" and the "known issues", which are part of the
system.

This commit was cherry-picked from master into a hotfix branch.