Show alert message when viewing content with different language
When user is viewing course content in different language than set in
their preferences, show an alert message with a button to change the
language back to their preferred one.
For students (aka users who are enrolled on the course), there is also
a button to change the current language as their preferred language for
this course.
For course staff (users who are not enrolled), there is a text
indicating that they can change the preferred language from their
profile.