Remove questionnaire feedback modal
On submission, questionnaires used to present a modal
with feedback on the progress of submission. And upon
acceptance or rejection the points badge and
the status of submission.

This removes that modal, and adds an overlay on top of the
the exercise to be displayed while the exercise is being posted.

Once the exercise is posted, or alternatively there is an error
in the submission, the overlay is removed and error message or
the graded form is being presented.