Fix missing quiz overlay
The quiz overlay was missing in the case when you have previously
submitted to the quiz and reload the chapter page. The quiz then
shows the previous feedback. When you submit again, the quiz
overlay was missing in this case. The "posting submission" progress
bar was not shown. This commit fixes this case.

The active elements code in chapter.js required a small change
because the code looks for .alert-danger elements in the feedback
HTML. Due to the quiz overlay fix, there is now always a new hidden
.alert-danger element. The fix for active elements is to ignore
hidden alerts.