Fix relative links in content loaded from the exercise service
This commit addresses many issues in the transformations of
the relative links.

Specifically, this commit fixes links from learning objects to
chapters and to static files, such as images in the course
repository.

* Links to chapters that are defined in subdirectories under
  the module directory in the RST course repository.
  The chapter key in the A+ URL contains the subdirectory names.
* Multilingual courses have language suffixes in RST chapters
  (chapter1_en.rst). The language suffixes are removed from
  the URLs since A+ URLs only contain the chapter key (that is
  the same for all language versions).
* Static file links in exercise descriptions when the exercise
  is defined in a chapter that is defined in a subdirectory.

Fixes https://github.com/apluslms/a-plus/issues/509

A-plus-rst-tools is updated at the same time so that the links
work locally in the HTML files under the course _build/html
directory. A+ transforms the links so that they work in A+.

Closes https://github.com/apluslms/a-plus/pull/463
The pull request did not work in all cases.

Note: https://github.com/apluslms/a-plus/issues/468
is still open and this commit does not fix it.
(Links to chapters that are a child of another chapter.)