Remove line numbers from the translation django.po file
The line numbers create a lot of noise in git diff.
It will be easier to review changes to translations when
the generated django.po file does not contain a lot of changed
line numbers in the comment lines.

Part of issue https://github.com/apluslms/a-plus/issues/676

The django.po file must be generated with these parameters so that
the line numbers are not added back again:
python manage.py makemessages --no-obsolete --add-location file