Update the EU-U.S. Privacy Shield description in the privacy warnings
The European Court of Justice has declared that
the EU-U.S. Privacy Shield is invalid. The Privacy Shield setting
should not be used in the LinkService destination_region field
in the future. The destination region of current services should be
updated in production. This commit leaves the Privacy Shield
value to the model so that nothing breaks immediately in production.
This commit updates the descriptions in the user interface so that
they would not make false promises about Privacy Shield.

Fixes https://github.com/apluslms/a-plus/issues/680