Redirect url without instance to the last course instance
Currently the o1 course uses a course url without any particular
instance. This URL allows the teacher to reference the course in
different websites, and papers. The concept behind this URL without any
instance is to have one immutable URL that can be shared without worrying
about the current instance of the course because the URL wil be
redirecte to the last instance of the course.

This commit allows to redirect any course URL that does not contain an
specific instance to the last course instance.