Fix issue regarding <a> tags introduced in 2b5c311
In the 2b5c311 commit an event handler was added to the 'a' tags to
control the navigability of the course by using the 'hl' query parameter
in the URL. However, the event handler caused several problems in some
courses that uses JS in their content. The event handler also caused
problems with the modal windows used in UI of A+. Therefore, in this
commit we remove the problematic code. However, it is important to
continue working on the Pull Request #736 to preserve the `hl` query
parameter when the user navigates through the course content.

Fixes #745