Remove target="_blank" from submitted files
With commit 5703158 all `<a target="_blank">` tags that open in new
tabs were adorned with the `glyphicon-new-window` icon. This icon is
used to indicate to the user that the link will open in a new tab.
However, this icon is also added to the "Submitted files" list, in the
"Inspect Files" view. This happens because the link that displays the
submitted files in a modal window has the `target="_blank"` attribute on
it. This commit removes that attribute as it is neither necessary nor
semantically correct.

Fixes #732