Modify actions when users respond to the prompt about lang settings
* If the user decides to continue browsing the course in a secondary
language. The only action we need to take is to redirect the user to a
new page that includes the secondary language query parameter in the
URL. We achieve this by using an <a> tag that contains the secondary
language query parameter in the href attribute.
* If the user decides to change the enrollment language, we must submit
a form with the new language, and then redirect the user to a page with
the new language query parameter.