Add multi-organization support to Haka login.
Adds organization attribute to UserProfile that is set based on information
received from Haka. Local organization of the A+ installation is specified
in settings, and used in determining whether a user is internal or external
to organization where A+ is installed. The button text for Haka login is
changed (the old text strings are still kept in settings for possible
future use). Closes #788.