Catch HTTP errors in the aplus-json course import
In the A+ edit course page, you can import the course from
the MOOC-Grader (aplus-json). Previously, if the HTTP GET for
the URL resulted in HTTP error response such as 404, the code
still tried to parse the response as JSON. That usually caused
JSON parser errors that were reported to the user (teacher) as
JSON parser errors. Now, the 4xx and 5xx errors result in
a clearer error message to the user (teacher).

In addition, use format_lazy with lazy strings instead of
str.format().