Redirect to enrollment when user is not enrolled
The course instance front page already redirected the user to the
enrollment page if the user was not enrolled in the course. Other pages
(e.g. exercises) didn't have this behavior, and instead displayed a 403
Forbidden error. This commit adds the redirect behavior to all course
pages except for teacher-specific ones.

Fixes #720