Add a group size indicator to exercise frame
Currently, the minimum and maximum allowed group sizes of an exercise
are not displayed in the exercise UI (except for a warning message, if
the user is not currently submitting with a group of appropriate size).
This commit adds a group size indicator to the header of the exercise
frame next to the deadline indicator.

Related to #808