Fix long language codes in JavaScript translations
The translation JSON files use only two-character language codes
such as "fi". However, the JavaScript code used to take the whole
lang attribute value from the <html> element. Under some circumstances,
the value is longer such as "en-gb". There are no translations
for the longer language codes and the defaut language is also only
"en". Therefore, it is best to take only the two first characters
from the html lang attribute value and use that as the language
for the translations.