Add alternative SISU format to results view
Adds a new checkbox in course results view that changes to
SISU-compatible output format. When exported as CSV, this should
be compatible with that is accepted by SISU. The SISU format has
fields that are currently not used, such as course grade and
assessment date, but teacher should fill these e.g. in Excel or
by some other tool.