Add translations to models and fields
Add verbose names and labels to models and fields, as well as their
translations. These are displayed on forms and in the admin view.
Also add some help texts to some fields.
Line changes and trailing commas are also added to many fields to help
with version control and be more in line with the style guide. Some
lines are also swapped around so there is more consistency in the order
of parameters.