Remove student self-unenrollment
The ability for students to unenroll themselves from a course turned out
to be an undesired feature. Therefore this commit disables it. UI
elements have been commented out, the /enroll/ URL has been commented
out, and the DELETE operation on the course students API view is now
only available for course staff.