Add a customizable note about personal data to the login page
Due to GDPR, users should be alerted before logging in about
how long personal data is stored in the system. The details may
vary in different organizations and thus, this note is
customizable in the server's local_settings.py. The structure of
the privacy notice may also vary in organizations and the note
in the login page may need to refer to certain sections of
the privacy notice.