Bypass submission limit in enrollment exercises
Allow unenrolled users to submit enrollment exercises regardless of
submission limits. This makes it possible for users to get back into a
course if they were accidentally removed by course staff, and the
enrollment exercise has a low submission limit.

In addition, the user's group selection now is ignored if they are
unenrolled. This allows the user to resubmit the enrollment exercise
even if they had activated a group before being removed from the course.

Fixes #864