Fix the <html> lang attribute value
Similarly, fix the language in the <body> class attribute
"lang-CODE".

Previously, the lang attribute matched the language set in
the user profile, but course pages may be displayed in other
languages. The lang attribute must match the language used in
the document. Multilingual courses set the user's language choice
in the enrollment object. It may be different from the user
profile's language.

templates/base.html outputs the lang attributes from the variable
request.LANGUAGE_CODE. The variable is set in lib/middleware.py
(class LocaleMiddleware), but the middleware has no access to
the course context and it can not easily determine the user's
language set in the course enrollment object. Therefore,
the request.LANGUAGE_CODE is overridden in the mixin class
that all course pages use.

Fixes #870