Avoid long language codes with regions in translations
Short language codes "en" and "fi" have been used in
Userprofile.language and Enrollment.language. However, if those
values are empty, Django translations define the language from
the HTTP request. The language from the request may be in the long
format "en-gb" with the region. However, this is inconsistent.
This commit removes the regional part so that short language codes
are always used. We have assumed in many places that only short
language codes are used and the long format breaks them.

The inconsistency of the language codes causes issues in at least
the following:
* <html lang> attribute (some courses use it in JavaScript code)
* grader protocol lang parameter (MOOC-Grader fails to load
  English content for the code "en-gb")
* translating exercise names in the A+ course front page
  (A+ database stores multilingual exercise names with the keys
  "en" and "fi" and it does not match "en-gb" with the English
  name.)