Fix the language of chapter exercises when using hl URL parameter
In addition, this fixes the language in modal links, such as
the submission feedback dialog in chapter pages.

When the user's enrollment.language differs from the language set
by the hl URL parameter, the chapter content was displayed in
the hl language, but the embedded exercises in the enrollment
language (selected in the course language menu).
This is fixed by adding the hl parameter to the AJAX requests
that load the embedded exercises.