Copy only the hl parameter in the link click handler
The click event handler was originally added for copying
the hl parameter to the next page. It was never necessary or
intentional to copy all query parameters. There are query parameters
that should not be copied to the subsequent pages, at least
the next parameter in the Django user login.

Fixes #745