Emphasize the selected languages in the course language warning
Add the names of the currently selected languages to the warning about
the selected course language (the box "You are browsing this course in
a different language..."). The warning now shows which language is
selected in the preferences (saved in the course enrollment or
user profile) and in which language the content is displayed.

The warning is displayed when the hl URL query parameter uses
a different language than is saved in the preferences.