Add filter types to AplusTableFilter
This commit adds support for customizing how each column is filtered in
a table marked with "filtered-table". The "data-filter-type" attribute
of the column determines the filtering type. The "none" type removes
filtering from the column altogether. The "options" type displays a set
of predefined filter options as checkboxes (the options are provided
in the "data-filter-options" attribute separated by vertical pipes).
When the type attribute is left unspecified, the old textbox filtering
is used.

Related to #831